# docs/swagger.yaml
basePath: /
definitions:
  handler.ShortenInputBody:
    properties:
      exp:
        minimum: 300
        type: integer
      url:
        type: string
    required:
    - exp
    - url
    type: object
  user.registerInput:
    properties:
      display_name:
        type: string
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - display_name
    - email
    - password
    - username
    type: object
info:
  contact: {}
  description: This is a health check API server.
  title: Health Check API
  version: "1.1"
paths:
  /health-check:
    get:
      consumes:
      - application/json
      description: Returns the health status of the service
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check endpoint
      tags:
      - health
  /v1/links/redirect/{code}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Shorten code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Found
      tags:
      - links
  /v1/links/shorten:
    post:
      consumes:
      - application/json
      description: Creates a short code for a given URL
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ShortenInputBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Shorten a URL
      tags:
      - links
  /v1/users/register:
    post:
      consumes:
      - application/json
      description: Register a new user with the provided information
      parameters:
      - description: User registration details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.registerInput'
      produces:
      - application/json
      responses: {}
      summary: Register a new user
      tags:
      - User
swagger: "2.0"